requirements.txt
gspread==3.7.0
oauth2client==4.1.3
python-dotenv==1.0.0
slack_sdk==3.19.2
google-api-python-client==2.66.0
